# perf(vector): Add heap to neighbour edges