--- a/posting/oracle.go
+++ b/posting/oracle.go
@@ -103,9 +103,11 @@ func (vt *viTxn) GetWithLockHeld(key []byte) (rval index.Value, rerr error) {
 }
 
 func (vt *viTxn) GetValueFromPostingList(pl *List) (rval index.Value, rerr error) {
-	val, err := pl.ValueWithLockHeld(vt.delegate.StartTs)
-	rval = val.Value
-	return rval, err
+	value := pl.findStaticValue(vt.delegate.StartTs)
+	if value == nil {
+		return nil, nil
+	}
+	return value.Postings[0].Value, nil
 }
 
 func (vt *viTxn) AddMutation(ctx context.Context, key []byte, t *index.KeyValue) error {