--- a/query/query.go
+++ b/query/query.go
@@ -1185,6 +1185,12 @@ func (sg *SubGraph) transformVars(doneVars map[string]varValue, path []*SubGraph
 			mt.Const = val
 			continue
 		}
+
+		if val, ok := newMap[0]; ok && len(newMap) == 1 {
+			mt.Const = val
+			continue
+		}
+
 		mt.Val = newMap
 	}
 	return nil