We now store the list of neighbours in a sorted distance from the starting node. This improves the accuracy and the time of the vector.
We are also using findStaticValue() function to get data faster.

Benchmarks:
1. Dataset 1 : 10k vector 100 dimension
  a. Time to ingest : 6 seconds
  b. Query time: 900us 
  c. Accuracy: 97.9%
  
2. Dataset 2: 1million vectors 1280 dimension
  a. Time to ingest: 7 mins
  b. Query time (without cache): 35 ms
  c. Query time (with cache): 5 ms